requirements.txt
pulsar-client==3.2.0
redis==4.5.5